+++ src/main/java/site/siredvin/gttruesteam/TrueSteamEntityTypeTags.java
@@ -1,10 +1,10 @@
 package site.siredvin.gttruesteam;
 
-import com.tterrag.registrate.providers.ProviderType;
-
 import net.minecraft.core.registries.Registries;
 import net.minecraft.tags.TagKey;
 import net.minecraft.world.entity.EntityType;
+
+import com.tterrag.registrate.providers.ProviderType;
 
 public class TrueSteamEntityTypeTags {
 
@@ -13,6 +13,7 @@ public class TrueSteamEntityTypeTags {
 
     public static void sayHi() {
         GTTrueSteam.REGISTRATE.addDataGenerator(ProviderType.ENTITY_TAGS, prov -> prov.addTag(NETHER)
-                .add(EntityType.BLAZE, EntityType.MAGMA_CUBE, EntityType.WITHER_SKELETON, EntityType.HOGLIN, EntityType.ZOGLIN, EntityType.PIGLIN, EntityType.PIGLIN_BRUTE));
+                .add(EntityType.BLAZE, EntityType.MAGMA_CUBE, EntityType.WITHER_SKELETON, EntityType.HOGLIN,
+                        EntityType.ZOGLIN, EntityType.PIGLIN, EntityType.PIGLIN_BRUTE));
     }
 }