+++ src/main/java/site/siredvin/gttruesteam/TrueSteamLang.java
@@ -2,6 +2,7 @@ package site.siredvin.gttruesteam;
 
 import net.minecraft.Util;
 import net.minecraft.network.chat.MutableComponent;
+
 import site.siredvin.gttruesteam.machines.spawner_extraction.MobType;
 
 public class TrueSteamLang {
@@ -67,12 +68,14 @@ public class TrueSteamLang {
 
     public static MutableComponent COIL_COOLING_CAPACITY = GTTrueSteam.REGISTRATE.addRawLang(COIL_COOLING_CAPACITY_KEY,
             "§bBase Cooling Capacity: §f%d ηK");
-    public static MutableComponent COOLING_COIL_RCC = GTTrueSteam.REGISTRATE.addLang("tooltip", GTTrueSteam.id("cooling_coil_rcc"),
+    public static MutableComponent COOLING_COIL_RCC = GTTrueSteam.REGISTRATE.addLang("tooltip",
+            GTTrueSteam.id("cooling_coil_rcc"),
             "§6Regulared cryo chamber:");
     public static MutableComponent COIL_COOLING_REDUCTION = GTTrueSteam.REGISTRATE.addRawLang(
             COIL_COOLING_REDUCTION_KEY,
             "    §bMax duration reduction: §f%d%%");
-    public static MutableComponent COOLING_PERFECT_OC = GTTrueSteam.REGISTRATE.addRawLang(COIL_COOLING_PERFECT_OC_KEY, "    §bPerfect OC: §fRecipe with %s or less base voltage");
+    public static MutableComponent COOLING_PERFECT_OC = GTTrueSteam.REGISTRATE.addRawLang(COIL_COOLING_PERFECT_OC_KEY,
+            "    §bPerfect OC: §fRecipe with %s or less base voltage");
     public static MutableComponent CHARGING_CYCLES = GTTrueSteam.REGISTRATE.addRawLang(CHARGING_CYCLES_KEY,
             "Charging for %d cycles");
     public static MutableComponent INFERNAL_CYCLES_UNTIL_THROTTLE = GTTrueSteam.REGISTRATE.addRawLang(