+++ src/main/java/site/siredvin/gttruesteam/util/RMHelpers.java
@@ -4,6 +4,7 @@ import com.gregtechceu.gtceu.api.GTValues;
 import com.gregtechceu.gtceu.api.recipe.GTRecipe;
 
 public class RMHelpers {
+
     public static int calculateOC(GTRecipe recipe) {
         var euT = recipe.getInputEUt().voltage();
         for (int i = GTValues.V.length - 1; i > -1; i--) {