+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Better purified infernal dust recipes
 - Buff boilers a little and introduce tier 3 for boilers
 - Spawner extraction machine
+- KubeJS plugin for generating new cooling coils
 
 ### Changed